Fix relationships with no data/guid
The data object for quota is typically null (see example in API docs
https://v3-apidocs.cloudfoundry.org/version/3.99.0/#the-space-object).
This ends up with the relationships dict containing the `quota` key but
the value is None so `dict.get('quota', {})` returns None.

Wrapping in braces and using `or` means that we will default to `{}`
when either the key doesn't exist (old behaviour) and also when the
value of the key is `None` (new behaviour) so the next `.get('guid')` is
safe.

Also removed the default of `None` from the `.get('guid')` since `None`
is the default already